requirements.txt
segpy
click
numpy
pillow
exit_codes
euclidian
toml